+++ src/main/java/br/com/zalf/prolog/webservice/commons/imagens/UploadImageHelper.java
@@ -7,6 +7,7 @@ import com.google.common.io.Files;
 import org.apache.commons.io.IOUtils;
 import org.jetbrains.annotations.NotNull;
 
+import javax.imageio.ImageIO;
 import java.io.File;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -19,11 +20,11 @@ import java.io.InputStream;
  */
 public class UploadImageHelper {
 
-    public static ImagemProLog uploadImagem(@NotNull final ImagemProLog imagemProLog,
-                                            @NotNull final InputStream inputStream,
+    public static ImagemProLog uploadImagem(@NotNull final InputStream inputStream,
                                             @NotNull final String amazonBucket)
-            throws IOException, S3FileSender.S3FileSenderException {
+            throws IOException, S3FileSender.S3FileSenderException, FileFormatNotSupportException {
 
+        final ImagemProLog imagemProLog = new ImagemProLog();
         final S3FileSender fileSender = new S3FileSender(
                 AmazonConstants.AWS_ACCESS_KEY_ID,
                 AmazonConstants.AWS_SECRET_KEY);
@@ -31,9 +32,8 @@ public class UploadImageHelper {
         // Pasta temporária da JVM
         final File tmpDir = Files.createTempDir();
         final File imageFile = createImageFile(tmpDir, inputStream, imageName);
-        final String mimeType = java.nio.file.Files.probeContentType(imageFile.toPath());
-        if (mimeType == null || !mimeType.split("/")[0].equals("image")) {
-            throw new IllegalArgumentException("Arquivo recebido não é uma imagem");
+        if (ImageIO.read(imageFile) == null) {
+            throw new FileFormatNotSupportException("O arquivo precisa ser uma imagem");
         }
         fileSender.sendFile(amazonBucket, imageName, imageFile);
         imagemProLog.setUrlImagem(fileSender.generateFileUrl(amazonBucket, imageName));
@@ -53,7 +53,7 @@ public class UploadImageHelper {
         final File file = new File(directory, imageName);
         final FileOutputStream outputStream = new FileOutputStream(file);
         IOUtils.copy(inputStream, outputStream);
-        IOUtils.closeQuietly(outputStream);
+        IOUtils.closeQuietly(outputStream, inputStream);
         return file;
     }
 }