+++ src/main/java/br/com/zalf/prolog/webservice/frota/checklist/modelo/ChecklistModeloResource.java
@@ -1,15 +1,12 @@
 package br.com.zalf.prolog.webservice.frota.checklist.modelo;
 
 import br.com.zalf.prolog.webservice.commons.imagens.Galeria;
-import br.com.zalf.prolog.webservice.commons.imagens.ImagemProLog;
 import br.com.zalf.prolog.webservice.commons.network.AbstractResponse;
 import br.com.zalf.prolog.webservice.commons.network.Response;
-import br.com.zalf.prolog.webservice.commons.network.ResponseWithCod;
 import br.com.zalf.prolog.webservice.frota.checklist.model.PerguntaRespostaChecklist;
 import br.com.zalf.prolog.webservice.interceptors.auth.Secured;
 import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
 import com.google.common.base.Preconditions;
-import org.glassfish.jersey.media.multipart.FormDataBodyPart;
 import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import org.glassfish.jersey.media.multipart.FormDataParam;
 
@@ -28,11 +25,8 @@ public class ChecklistModeloResource {
     @POST
     @Secured(permissions = {Pilares.Frota.Checklist.Modelo.ALTERAR, Pilares.Frota.Checklist.Modelo.CADASTRAR})
     public Response insertModeloChecklist(ModeloChecklist modeloChecklist) {
-        if (service.insertModeloChecklist(modeloChecklist)) {
-            return Response.ok("Modelo de checklist inserido com sucesso");
-        } else {
-            return Response.error("Erro ao inserir modelo de checklist");
-        }
+        service.insertModeloChecklist(modeloChecklist);
+        return Response.ok("Modelo de checklist inserido com sucesso");
     }
 
     @GET
@@ -115,21 +109,9 @@ public class ChecklistModeloResource {
     @Consumes({MediaType.MULTIPART_FORM_DATA})
     public AbstractResponse insertImagemGaleria(@PathParam("codEmpresa") Long codEmpresa,
                                                 @FormDataParam("file") InputStream fileInputStream,
-                                                @FormDataParam("file") FormDataContentDisposition fileDetail,
-                                                @FormDataParam("imagem") FormDataBodyPart jsonPart) {
+                                                @FormDataParam("file") FormDataContentDisposition fileDetail) {
         Preconditions.checkNotNull(codEmpresa, "Código da empresa não pode ser null!");
 
-        jsonPart.setMediaType(MediaType.APPLICATION_JSON_TYPE);
-        final ImagemProLog imagemProLog = jsonPart.getValueAs(ImagemProLog.class);
-        if (imagemProLog == null) {
-            return Response.error("ERRO! Imagem veio nula");
-        } else {
-            final Long codImagem = service.insertImagem(codEmpresa, fileInputStream, imagemProLog);
-            if (codImagem != null) {
-                return ResponseWithCod.ok("Imagem inserida com sucesso", codImagem);
-            } else {
-                return Response.error("Erro ao inserir imagem");
-            }
-        }
+        return service.insertImagem(codEmpresa, fileInputStream);
     }
 }