+++ src/main/java/br/com/zalf/prolog/webservice/gente/controleintervalo/ControleIntervaloService.java
@@ -9,7 +9,7 @@ import br.com.zalf.prolog.webservice.commons.network.ResponseWithCod;
 import br.com.zalf.prolog.webservice.commons.util.Log;
 import br.com.zalf.prolog.webservice.gente.controleintervalo.model.*;
 import br.com.zalf.prolog.webservice.permissao.pilares.Pilares;
-import com.sun.istack.internal.NotNull;
+import org.jetbrains.annotations.NotNull;
 
 import java.sql.SQLException;
 import java.util.List;
@@ -30,7 +30,18 @@ public class ControleIntervaloService {
             Log.e(TAG, String.format("Erro ao buscar os tipos de intervalos. \n" +
                     "codUnidade: %d \n" +
                     "withCargos: %b", codUnidade, withCargos), e);
-            return null;
+            throw new RuntimeException(e);
+        }
+    }
+
+    public TipoIntervalo getTipoIntervalo(Long codUnidade, Long codTipoIntervalo) {
+        try {
+            return dao.getTipoIntervalo(codUnidade, codTipoIntervalo);
+        } catch (SQLException e) {
+            Log.e(TAG, String.format("Erro ao buscar tipo de intervalo. \n" +
+                    "codUnidade: %d \n" +
+                    "codTipoIntervalo: %d", codUnidade, codTipoIntervalo), e);
+            throw new RuntimeException(e);
         }
     }
 
@@ -74,13 +85,15 @@ public class ControleIntervaloService {
                     "codTipo: %s \n" +
                     "limit: %d \n" +
                     "offset: %d", cpf, codTipo, limit, offset), e);
-            return null;
+            throw new RuntimeException(e);
         }
     }
 
-    public AbstractResponse insertTipoIntervalo(@NotNull TipoIntervalo tipoIntervalo) {
+    public AbstractResponse insertTipoIntervalo(@NotNull final TipoIntervalo tipoIntervalo) {
         try {
-            return ResponseWithCod.ok("Intervalo inserido com sucesso", dao.insertTipoIntervalo(tipoIntervalo,
+            return ResponseWithCod.ok(
+                    "Tipo de intervalo inserido com sucesso",
+                    dao.insertTipoIntervalo(tipoIntervalo,
                     Injection.provideDadosIntervaloChangedListener()));
         } catch (Throwable e){
             Log.e(TAG, "Erro ao inserir o tipo de intervalo", e);
@@ -172,7 +185,7 @@ public class ControleIntervaloService {
                     "codUnidade: %d \n" +
                     "cpf: %d \n" +
                     "codTipo: %d", codUnidade, cpf, codTipo), e);
-            return null;
+            throw new RuntimeException(e);
         }
     }